server & web & config: server- Change the ws to include a new filter doses, two modes checkbox and input, web create a new doses filter. config- change the cellbase ws because ws.babelomics is actualizating DB.

+++ cnvs-lib/src/main/java/org/babelomics/cnvs/lib/models/CNV.java
@@ -47,7 +47,7 @@ public class CNV {
     @Property("t")
     private int type;
     @Property("d")
-    private float doses;
+    private double doses;
     @Property("cs")
     private int clinicalSig;
     @Property("i")
